+++ src/ipyelk/contrib/molds/__init__.py
@@ -1,2 +1,3 @@
+"""Reusable shape factories for ``ipyelk``."""
 # Copyright (c) 2024 ipyelk contributors.
 # Distributed under the terms of the Modified BSD License.